+++ client/interop-mobile-multiplatform/composeApp/src/desktopMain/kotlin/org/calypsonet/keyple/demo/reload/remote/Datastore.kt
@@ -1,0 +1,16 @@
+/* ******************************************************************************
+ * Copyright (c) 2024 Calypso Networks Association https://calypsonet.org/
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the BSD 3-Clause License which is available at
+ * https://opensource.org/licenses/BSD-3-Clause.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ ****************************************************************************** */
+package org.calypsonet.keyple.demo.reload.remote
+
+actual class DataStorePathProducer {
+  actual fun producePath(): String {
+    return dataStoreFileName
+  }
+}