+++ server/src/main/java/org/calypsonet/keyple/demo/reload/remote/server/card/CardController.java
@@ -23,8 +23,21 @@ import org.eclipse.keyple.distributed.SyncNodeServer;
 @Path("/card")
 public class CardController {
 
-  @Inject CardConfigurator cardConfigurator;
+  @Inject CardRepository cardRepository;
   @Inject CardSamObserver cardSamObserver;
+
+  /**
+   * Returns the exported card selection scenario, as a JSON string.
+   *
+   * @return A JSON string containing the exported card selection scenario.
+   */
+  @GET
+  @Path("/export-card-selection-scenario")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response exportCardSelectionScenario() {
+    String cardSelectionScenarioJsonString = cardRepository.exportCardSelectionScenario();
+    return Response.ok(cardSelectionScenarioJsonString).build();
+  }
 
   /**
    * The endpoint access associated with the remote plugin server.