+++ server/src/main/java/org/calypsonet/keyple/demo/reload/remote/server/card/CardService.java
@@ -30,6 +30,7 @@ import org.eclipse.keyple.core.service.resource.CardResource;
 import org.eclipse.keyple.core.service.resource.CardResourceServiceProvider;
 import org.eclipse.keyple.core.util.HexUtil;
 import org.eclipse.keypop.calypso.card.card.CalypsoCard;
+import org.eclipse.keypop.calypso.card.transaction.CardIOException;
 import org.eclipse.keypop.reader.CardReader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -70,6 +71,8 @@ public class CardService {
   private static final String CARD_NOT_PERSONALIZED = "Card not personalized.";
   private static final String ENVIRONMENT_EXPIRED = "Environment expired.";
   private static final String RUNTIME_EXCEPTION = "Runtime exception: ";
+  private static final String PROCESSED_CARD_SELECTION_SCENARIO_JSON_STRING =
+      "processedCardSelectionScenarioJsonString";
 
   private final DateTimeFormatter dateTimeFormatter =
       DateTimeFormatter.ofPattern("d MMMM yyyy", Locale.ENGLISH);
@@ -251,7 +254,7 @@ public class CardService {
 
     if (!CardConstant.Companion.getALLOWED_FILE_STRUCTURES()
         .contains(calypsoCard.getApplicationSubtype())) {
-      return new AnalyzeContractsOutputDto(Collections.emptyList(), 2);
+      return new AnalyzeContractsOutputDto(Collections.emptyList(), 3);
     }
 
     CardResource samResource =
@@ -276,7 +279,7 @@ public class CardService {
               .setStatus(FAIL)
               .setType(SECURED_READ)
               .setCardSerialNumber(appSerialNumber));
-      return new AnalyzeContractsOutputDto(Collections.emptyList(), 3);
+      return new AnalyzeContractsOutputDto(Collections.emptyList(), 4);
     } catch (ExpiredEnvironmentException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_ANALYZING_THE_CONTRACTS, e.getMessage());
       activityService.push(
@@ -285,8 +288,8 @@ public class CardService {
               .setStatus(FAIL)
               .setType(SECURED_READ)
               .setCardSerialNumber(appSerialNumber));
-      return new AnalyzeContractsOutputDto(Collections.emptyList(), 4);
-    } catch (RuntimeException e) {
+      return new AnalyzeContractsOutputDto(Collections.emptyList(), 5);
+    } catch (CardIOException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_ANALYZING_THE_CONTRACTS, e.getMessage(), e);
       activityService.push(
           new Activity()
@@ -295,6 +298,15 @@ public class CardService {
               .setType(SECURED_READ)
               .setCardSerialNumber(appSerialNumber));
       return new AnalyzeContractsOutputDto(Collections.emptyList(), 1);
+    } catch (RuntimeException e) {
+      logger.error(AN_ERROR_OCCURRED_WHILE_ANALYZING_THE_CONTRACTS, e.getMessage(), e);
+      activityService.push(
+          new Activity()
+              .setPlugin(pluginType)
+              .setStatus(FAIL)
+              .setType(SECURED_READ)
+              .setCardSerialNumber(appSerialNumber));
+      return new AnalyzeContractsOutputDto(Collections.emptyList(), 2);
     } finally {
       CardResourceServiceProvider.getService().releaseCardResource(samResource);
     }
@@ -308,7 +320,7 @@ public class CardService {
 
     if (!CardConstant.Companion.getALLOWED_FILE_STRUCTURES()
         .contains(calypsoCard.getApplicationSubtype())) {
-      return new WriteContractOutputDto(2);
+      return new WriteContractOutputDto(3);
     }
 
     logger.info("Inserted card application serial number: {}", appSerialNumber);
@@ -320,7 +332,7 @@ public class CardService {
       Card card = cardRepository.readCard(cardReader, calypsoCard, samResource);
       if (card == null) {
         // If card has not been read previously, throw error
-        return new WriteContractOutputDto(3);
+        return new WriteContractOutputDto(4);
       }
       // logger.info("{}", card); deactivate until LocalDate is properly processed by KeypleUtil
       insertNewContract(inputData.getContractTariff(), inputData.getTicketToLoad(), card);
@@ -337,7 +349,7 @@ public class CardService {
                           ? ": " + inputData.getTicketToLoad()
                           : "")));
       return new WriteContractOutputDto(statusCode);
-    } catch (RuntimeException e) {
+    } catch (CardIOException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_WRITING_THE_CONTRACT, e.getMessage(), e);
       activityService.push(
           new Activity()
@@ -347,6 +359,16 @@ public class CardService {
               .setCardSerialNumber(appSerialNumber)
               .setContractLoaded(""));
       return new WriteContractOutputDto(1);
+    } catch (RuntimeException e) {
+      logger.error(AN_ERROR_OCCURRED_WHILE_WRITING_THE_CONTRACT, e.getMessage(), e);
+      activityService.push(
+          new Activity()
+              .setPlugin(pluginType)
+              .setStatus(FAIL)
+              .setType(RELOAD)
+              .setCardSerialNumber(appSerialNumber)
+              .setContractLoaded(""));
+      return new WriteContractOutputDto(2);
     } finally {
       CardResourceServiceProvider.getService().releaseCardResource(samResource);
     }
@@ -360,7 +382,7 @@ public class CardService {
 
     if (!CardConstant.Companion.getALLOWED_FILE_STRUCTURES()
         .contains(calypsoCard.getApplicationSubtype())) {
-      return new CardIssuanceOutputDto(2);
+      return new CardIssuanceOutputDto(3);
     }
 
     CardResource samResource =
@@ -375,7 +397,7 @@ public class CardService {
               .setType(ISSUANCE)
               .setCardSerialNumber(appSerialNumber));
       return new CardIssuanceOutputDto(0);
-    } catch (RuntimeException e) {
+    } catch (CardIOException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_INITIALIZING_THE_CARD, e.getMessage(), e);
       activityService.push(
           new Activity()
@@ -384,21 +406,42 @@ public class CardService {
               .setType(ISSUANCE)
               .setCardSerialNumber(appSerialNumber));
       return new CardIssuanceOutputDto(1);
+    } catch (RuntimeException e) {
+      logger.error(AN_ERROR_OCCURRED_WHILE_INITIALIZING_THE_CARD, e.getMessage(), e);
+      activityService.push(
+          new Activity()
+              .setPlugin(pluginType)
+              .setStatus(FAIL)
+              .setType(ISSUANCE)
+              .setCardSerialNumber(appSerialNumber));
+      return new CardIssuanceOutputDto(2);
     } finally {
       CardResourceServiceProvider.getService().releaseCardResource(samResource);
     }
   }
 
   SelectAppAndAnalyzeContractsOutputDto selectAppAndAnalyzeContracts(
-      CardReader cardReader, SelectAppAndAnalyzeContractsInputDto inputData) {
+      CardReader cardReader,
+      SelectAppAndAnalyzeContractsInputDto inputData,
+      Properties properties) {
 
     String pluginType = inputData.getPluginType();
+    String processedCardSelectionScenarioJsonString =
+        properties != null
+            ? properties.getProperty(PROCESSED_CARD_SELECTION_SCENARIO_JSON_STRING)
+            : null;
 
     // Select application
     CalypsoCard calypsoCard;
     try {
-      calypsoCard = cardRepository.selectCard(cardReader);
-    } catch (RuntimeException e) {
+      if (processedCardSelectionScenarioJsonString != null) {
+        calypsoCard =
+            cardRepository.importProcessedCardSelectionScenario(
+                processedCardSelectionScenarioJsonString);
+      } else {
+        calypsoCard = cardRepository.selectCard(cardReader);
+      }
+    } catch (CardIOException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_ANALYZING_THE_CONTRACTS, e.getMessage(), e);
       activityService.push(
           new Activity()
@@ -408,6 +451,16 @@ public class CardService {
               .setCardSerialNumber(""));
       return new SelectAppAndAnalyzeContractsOutputDto(
           "", Collections.emptyList(), 1, e.getMessage());
+    } catch (RuntimeException e) {
+      logger.error(AN_ERROR_OCCURRED_WHILE_ANALYZING_THE_CONTRACTS, e.getMessage(), e);
+      activityService.push(
+          new Activity()
+              .setPlugin(pluginType)
+              .setStatus(FAIL)
+              .setType(SECURED_READ)
+              .setCardSerialNumber(""));
+      return new SelectAppAndAnalyzeContractsOutputDto(
+          "", Collections.emptyList(), 2, e.getMessage());
     }
 
     // Analyze contracts
@@ -492,18 +545,21 @@ public class CardService {
         message = "Success";
         break;
       case 1:
-        message = "Server error";
+        message = "Card communication error";
         break;
       case 2:
+        message = "Server error";
+        break;
+      case 3:
         message =
             "Invalid card\nFile structure "
                 + HexUtil.toHex(calypsoCard.getApplicationSubtype())
                 + "h not supported";
         break;
-      case 3:
+      case 4:
         message = "Environment error: wrong version number";
         break;
-      case 4:
+      case 5:
         message = "Environment error: end date expired";
         break;
       default:
@@ -514,15 +570,25 @@ public class CardService {
   }
 
   SelectAppAndLoadContractOutputDto selectAppAndLoadContract(
-      CardReader cardReader, SelectAppAndLoadContractInputDto inputData) {
+      CardReader cardReader, SelectAppAndLoadContractInputDto inputData, Properties properties) {
 
     String pluginType = inputData.getPluginType();
+    String processedCardSelectionScenarioJsonString =
+        properties != null
+            ? properties.getProperty(PROCESSED_CARD_SELECTION_SCENARIO_JSON_STRING)
+            : null;
 
     // Select application
     CalypsoCard calypsoCard;
     try {
-      calypsoCard = cardRepository.selectCard(cardReader);
-    } catch (RuntimeException e) {
+      if (processedCardSelectionScenarioJsonString != null) {
+        calypsoCard =
+            cardRepository.importProcessedCardSelectionScenario(
+                processedCardSelectionScenarioJsonString);
+      } else {
+        calypsoCard = cardRepository.selectCard(cardReader);
+      }
+    } catch (CardIOException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_WRITING_THE_CONTRACT, e.getMessage(), e);
       activityService.push(
           new Activity()
@@ -531,6 +597,15 @@ public class CardService {
               .setType(SECURED_READ)
               .setCardSerialNumber(""));
       return new SelectAppAndLoadContractOutputDto(1, e.getMessage());
+    } catch (RuntimeException e) {
+      logger.error(AN_ERROR_OCCURRED_WHILE_WRITING_THE_CONTRACT, e.getMessage(), e);
+      activityService.push(
+          new Activity()
+              .setPlugin(pluginType)
+              .setStatus(FAIL)
+              .setType(SECURED_READ)
+              .setCardSerialNumber(""));
+      return new SelectAppAndLoadContractOutputDto(2, e.getMessage());
     }
 
     String selectedApplicationSerialNumber =
@@ -557,9 +632,12 @@ public class CardService {
         message = "Success";
         break;
       case 1:
-        message = "Server error";
+        message = "Card communication error";
         break;
       case 2:
+        message = "Server error";
+        break;
+      case 3:
         message =
             "Invalid card\nFile structure "
                 + HexUtil.toHex(calypsoCard.getApplicationSubtype())
@@ -572,15 +650,25 @@ public class CardService {
   }
 
   SelectAppAndPersonalizeCardOutputDto selectAppAndPersonalizeCard(
-      CardReader cardReader, SelectAppAndPersonalizeCardInputDto inputData) {
+      CardReader cardReader, SelectAppAndPersonalizeCardInputDto inputData, Properties properties) {
 
     String pluginType = inputData.getPluginType();
+    String processedCardSelectionScenarioJsonString =
+        properties != null
+            ? properties.getProperty(PROCESSED_CARD_SELECTION_SCENARIO_JSON_STRING)
+            : null;
 
     // Select application
     CalypsoCard calypsoCard;
     try {
-      calypsoCard = cardRepository.selectCard(cardReader);
-    } catch (RuntimeException e) {
+      if (processedCardSelectionScenarioJsonString != null) {
+        calypsoCard =
+            cardRepository.importProcessedCardSelectionScenario(
+                processedCardSelectionScenarioJsonString);
+      } else {
+        calypsoCard = cardRepository.selectCard(cardReader);
+      }
+    } catch (CardIOException e) {
       logger.error(AN_ERROR_OCCURRED_WHILE_INITIALIZING_THE_CARD, e.getMessage(), e);
       activityService.push(
           new Activity()
@@ -589,6 +677,15 @@ public class CardService {
               .setType(SECURED_READ)
               .setCardSerialNumber(""));
       return new SelectAppAndPersonalizeCardOutputDto(1, e.getMessage());
+    } catch (RuntimeException e) {
+      logger.error(AN_ERROR_OCCURRED_WHILE_INITIALIZING_THE_CARD, e.getMessage(), e);
+      activityService.push(
+          new Activity()
+              .setPlugin(pluginType)
+              .setStatus(FAIL)
+              .setType(SECURED_READ)
+              .setCardSerialNumber(""));
+      return new SelectAppAndPersonalizeCardOutputDto(2, e.getMessage());
     }
 
     // Init card
@@ -604,9 +701,12 @@ public class CardService {
         message = "Success";
         break;
       case 1:
-        message = "Server error";
+        message = "Card communication error";
         break;
       case 2:
+        message = "Server error";
+        break;
+      case 3:
         message =
             "Invalid card\nFile structure "
                 + HexUtil.toHex(calypsoCard.getApplicationSubtype())